Fixes and cleanup
