+++ application/org.openjdk.jmc.feature.flightrecorder/feature.xml
@@ -114,6 +114,13 @@
          install-size="0"
          version="0.0.0"
          unpack="false"/>
+         
+   <plugin
+         id="org.openjdk.jmc.flightrecorder.serializers"
+         download-size="0"
+         install-size="0"
+         version="0.0.0"
+         unpack="false"/>
 
    <plugin
          id="org.openjdk.jmc.rjmx.services.jfr"
