+++ application/org.openjdk.jmc.flightrecorder.graphview/META-INF/MANIFEST.MF
@@ -3,9 +3,9 @@ Bundle-ManifestVersion: 2
 Bundle-Name: GraphView
 Bundle-SymbolicName: org.openjdk.jmc.flightrecorder.graphview;singleton:=true
 Bundle-Version: 8.1.0.qualifier
-Require-Bundle: org.openjdk.jmc.flightrecorder;bundle-version="8.1.0",
- org.openjdk.jmc.flightrecorder.ui;bundle-version="8.1.0",
- org.openjdk.jmc.flightrecorder.serializers;bundle-version="8.1.0",
+Require-Bundle: org.openjdk.jmc.flightrecorder,
+ org.openjdk.jmc.flightrecorder.ui,
+ org.openjdk.jmc.flightrecorder.serializers,
  org.junit,
  org.junit.jupiter.api,
  org.junit.jupiter.engine,
