--- application/tests/org.openjdk.jmc.flightrecorder.flameview.test/META-INF/MANIFEST.MF
@@ -1,23 +1,0 @@
-Manifest-Version: 1.0
-Bundle-Version: 8.1.0.qualifier
-Bundle-Name: Flameview Tests
-Bundle-ManifestVersion: 2
-Bundle-SymbolicName: org.openjdk.jmc.flightrecorder.flameview.test;singleton:=true
-Bundle-Vendor: Oracle Corporation
-Require-Bundle: org.junit,
- org.openjdk.jmc.flightrecorder,
- org.openjdk.jmc.flightrecorder.flameview,
- org.openjdk.jmc.common.test,
- org.openjdk.jmc.flightrecorder.test
-Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Bundle-ActivationPolicy: lazy
-Automatic-Module-Name: org.openjdk.jmc.flightrecorder.flameview.test
-Import-Package: org.openjdk.jmc.common,
- org.openjdk.jmc.common.item,
- org.openjdk.jmc.common.unit,
- org.openjdk.jmc.common.util,
- org.openjdk.jmc.flightrecorder,
- org.openjdk.jmc.flightrecorder.jdk,
- org.openjdk.jmc.flightrecorder.stacktrace,
- org.openjdk.jmc.flightrecorder.test.util
-
