+++ application/tests/pom.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--   
-   Copyright (c) 2018, 2020, Oracle and/or its affiliates. All rights reserved.
+   Copyright (c) 2018, 2021, Oracle and/or its affiliates. All rights reserved.
    
    DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
    
@@ -45,7 +45,6 @@
 		<module>org.openjdk.jmc.flightrecorder.controlpanel.ui.configuration.test</module>
 		<module>org.openjdk.jmc.flightrecorder.controlpanel.ui.test</module>
 		<module>org.openjdk.jmc.flightrecorder.ext.jfx.test</module>
-		<module>org.openjdk.jmc.flightrecorder.flameview.test</module>
 		<module>org.openjdk.jmc.flightrecorder.ui.test</module>
 		<module>org.openjdk.jmc.greychart.test</module>
 		<module>org.openjdk.jmc.ide.launch.test</module>
