+++ core/tests/org.openjdk.jmc.flightrecorder.serializers.test/src/test/resources/flamegraph-attribute-inverted.json
