+++ core/tests/org.openjdk.jmc.flightrecorder.serializers.test/src/test/resources/flamegraph-attribute.json
